Surfacing qpu headers in user code
