+++ python/runtime/cudaq/platform/py_alt_launch_kernel.cpp
@@ -10,6 +10,7 @@
 #include "common/AnalogHamiltonian.h"
 #include "common/ArgumentWrapper.h"
 #include "common/Environment.h"
+#include "common/Timing.h"
 #include "cudaq/Optimizer/Builder/Marshal.h"
 #include "cudaq/Optimizer/Builder/Runtime.h"
 #include "cudaq/Optimizer/CAPI/Dialects.h"
@@ -20,6 +21,7 @@
 #include "cudaq/platform.h"
 #include "cudaq/platform/nvqpp_interface.h"
 #include "cudaq/platform/qpu.h"
+#include "cudaq/runtime/logger/logger.h"
 #include "cudaq_internal/compiler/ArgumentConversion.h"
 #include "cudaq_internal/compiler/LayoutInfo.h"
 #include "cudaq_internal/compiler/TracePassInstrumentation.h"
