+++ python/tests/interop/test_cpp_quantum_algorithm_module.cpp
@@ -6,8 +6,7 @@
  * the terms of the Apache License 2.0 which accompanies this distribution.    *
  ******************************************************************************/
 
-#include "cudaq.h"
-#include "cudaq/algorithms/sample.h"
+#include "cudaq/qis/qkernel.h"
 #include "quantum_lib/quantum_lib.h"
 #include "runtime/interop/PythonCppInterop.h"
 #include <nanobind/nanobind.h>
