+++ python/utils/LinkedLibraryHolder.cpp
@@ -10,7 +10,7 @@
 #include "common/FmtCore.h"
 #include "common/PluginUtils.h"
 #include "cudaq/Support/TargetConfigYaml.h"
-#include "cudaq/platform.h"
+#include "cudaq/platform/quantum_platform.h"
 #include "cudaq/runtime/logger/logger.h"
 #include "cudaq/target_control.h"
 #include "nvqir/CircuitSimulator.h"
