+++ runtime/common/BaseRemoteRESTQPU.h
@@ -8,20 +8,21 @@
 
 #pragma once
 
-#include "NoiseModel.h"
 #include "common/Environment.h"
 #include "common/ExecutionContext.h"
 #include "common/Executor.h"
-#include "common/ExtraPayloadProvider.h"
+#include "common/KernelExecution.h"
 #include "common/Resources.h"
-#include "cudaq.h"
-#include "cudaq/platform.h"
+#include "cudaq/Support/TargetConfig.h"
+#include "cudaq/platform/platform_iface.h"
 #include "cudaq/platform/qpu.h"
 #include "cudaq/platform/qpu_utils.h"
 #include "cudaq/runtime/logger/logger.h"
+#include "cudaq/utils/cudaq_utils.h"
 #include "cudaq_internal/compiler/CompiledModuleHelper.h"
 #include "cudaq_internal/compiler/Compiler.h"
 #include "cudaq_internal/compiler/JIT.h"
+#include <filesystem>
 #include <fstream>
 #include <netinet/in.h>
 #include <sys/socket.h>
@@ -34,20 +35,10 @@ void setResourceCounts(cudaq::Resources &&);
 bool isUsingResourceCounterSimulator();
 } // namespace nvqir
 
-// When compiled into the Python extension, the LLVM Registry<T> Head/Tail
-// static-inline pointers have hidden visibility (local 'b' symbols) instead
-// of GNU-unique ('u') symbols. This means registry::get<T> in the Python
-// extension sees an empty list even though dlopen'd .so plugins registered
-// into libcudaq-common's unique-symbol registry.  These C-linkage helpers
-// perform the lookup inside libcudaq-common so it works across DSO boundaries.
-#ifdef CUDAQ_PYTHON_EXTENSION
-extern "C" cudaq::ServerHelper *cudaq_find_server_helper(const char *name);
-extern "C" bool cudaq_has_server_helper(const char *name);
-extern "C" cudaq::Executor *cudaq_find_executor(const char *name);
-extern "C" bool cudaq_has_executor(const char *name);
-#endif
-
 namespace cudaq {
+void set_random_seed(std::size_t seed);
+std::size_t get_random_seed();
+class noise_model;
 
 class BaseRemoteRESTQPU : public QPU {
 protected:
@@ -69,9 +60,9 @@ protected:
   // Pointer to the concrete Executor for this QPU
   std::unique_ptr<cudaq::Executor> executor;
 
-  /// @brief Pointer to the concrete ServerHelper, provides
+  /// @brief Pointer to the forward declared ServerHelper, provides
   /// specific JSON payloads and POST/GET URL paths.
-  std::unique_ptr<cudaq::ServerHelper> serverHelper;
+  cudaq::owning_ptr<cudaq::ServerHelper> serverHelper;
 
   /// @brief Mapping of general key-values for backend configuration.
   std::map<std::string, std::string> backendConfig;
@@ -221,41 +212,8 @@ public:
     // Set the qpu name
     qpuName = mutableBackend;
     // Create the ServerHelper for this QPU and give it the backend config
-#ifdef CUDAQ_PYTHON_EXTENSION
-    serverHelper.reset(cudaq_find_server_helper(qpuName.c_str()));
-#else
-    serverHelper = cudaq::registry::get<cudaq::ServerHelper>(qpuName);
-#endif
-    if (!serverHelper) {
-      throw std::runtime_error("ServerHelper not found for target: " + qpuName);
-    }
-
-    serverHelper->initialize(backendConfig);
-    CUDAQ_INFO("Retrieving executor with name {}", qpuName);
-#ifdef CUDAQ_PYTHON_EXTENSION
-    bool hasExecutor = cudaq_has_executor(qpuName.c_str());
-    CUDAQ_INFO("Is this executor registered? {}", hasExecutor);
-    executor = hasExecutor ? std::unique_ptr<cudaq::Executor>(
-                                 cudaq_find_executor(qpuName.c_str()))
-                           : std::make_unique<cudaq::Executor>();
-#else
-    CUDAQ_INFO("Is this executor registered? {}",
-               cudaq::registry::isRegistered<cudaq::Executor>(qpuName));
-    executor = cudaq::registry::isRegistered<cudaq::Executor>(qpuName)
-                   ? cudaq::registry::get<cudaq::Executor>(qpuName)
-                   : std::make_unique<cudaq::Executor>();
-#endif
-
-    // Give the server helper to the executor
-    executor->setServerHelper(serverHelper.get());
-
-    // Construct the runtime target
-    RuntimeTarget runtimeTarget;
-    runtimeTarget.config = targetConfig;
-    runtimeTarget.name = mutableBackend;
-    runtimeTarget.description = targetConfig.Description;
-    runtimeTarget.runtimeConfig = backendConfig;
-    serverHelper->setRuntimeTarget(runtimeTarget);
+    detail::initServerHelperAndExecutor(qpuName, backendConfig, targetConfig,
+                                        serverHelper, executor);
   }
 
   /// @brief Launch the kernel. Extract the Quake code and lower to the
@@ -275,13 +233,13 @@ public:
           "Remote rest execution can only be performed via cudaq::sample(), "
           "cudaq::observe(), cudaq::run(), or cudaq::contrib::draw().");
 
-    auto [module, context] = Compiler::loadQuakeCodeByName(kernelName);
+    auto [moduleOp, context] = Compiler::loadQuakeCodeByName(kernelName);
 
     // Get the Quake code, lowered according to config file.
     Compiler compiler(serverHelper.get(), backendConfig, targetConfig,
                       noiseModel, emulate);
     auto codes =
-        compiler.lowerQuakeCode(executionContext, kernelName, module, args);
+        compiler.lowerQuakeCode(executionContext, kernelName, moduleOp, args);
     completeLaunchKernel(kernelName, std::move(codes));
 
     // NB: Kernel should/will never return dynamic results.
@@ -335,7 +293,7 @@ public:
       cudaq::ExecutionContext context("tracer");
       context.executionManager = cudaq::getDefaultExecutionManager();
       assert(codes[0].jit);
-      cudaq::get_platform().with_execution_context(
+      cudaq::platform::with_execution_context(
           context, [&]() { codes[0].jit->run(kernelName); });
       executionContext->kernelTrace = std::move(context.kernelTrace);
       return;
@@ -346,7 +304,7 @@ public:
       context.executionManager = cudaq::getDefaultExecutionManager();
       assert(codes.size() == 1 && codes[0].jit && codes[0].resourceCounts);
       nvqir::setResourceCounts(std::move(codes[0].resourceCounts.value()));
-      cudaq::get_platform().with_execution_context(
+      cudaq::platform::with_execution_context(
           context, [&]() { codes[0].jit->run(kernelName); });
       return;
     }
@@ -426,7 +384,7 @@ public:
                     executionContext ? executionContext->warnedNamedMeasurements
                                      : false;
                 assert(codes[i].jit);
-                cudaq::get_platform().with_execution_context(
+                cudaq::platform::with_execution_context(
                     context, [&]() { codes[i].jit->run(kernelName); });
 
                 if (isObserve) {
