+++ runtime/common/CMakeLists.txt
@@ -20,6 +20,7 @@ set(COMMON_RUNTIME_SRC
   Executor.cpp
   ExtraPayloadProvider.cpp
   Future.cpp
+  KernelExecution.cpp
   NoiseModel.cpp
   RecordLogParser.cpp
   Resources.cpp
