+++ runtime/common/ServerHelper.cpp
@@ -7,45 +7,11 @@
  ******************************************************************************/
 
 #include "ServerHelper.h"
+#include "cudaq/platform/qpu_utils.h"
+#include "cudaq/utils/owning_ptr.h"
 #include "nlohmann/json.hpp"
 
 namespace cudaq {
-
-KernelExecution::KernelExecution(const std::string &n, const std::string &c,
-                                 std::optional<cudaq::JitEngine> jit,
-                                 std::optional<Resources> rc,
-                                 std::vector<std::size_t> &m)
-    : name(n), code(c), jit(jit), resourceCounts(rc),
-      output_names(nlohmann::json{}), mapping_reorder_idx(m),
-      user_data(nlohmann::json{}) {}
-KernelExecution::KernelExecution(const std::string &n, const std::string &c,
-                                 std::optional<cudaq::JitEngine> jit,
-                                 std::optional<Resources> rc, nlohmann::json &o,
-                                 std::vector<std::size_t> &m)
-    : name(n), code(c), jit(jit), resourceCounts(rc), output_names(o),
-      mapping_reorder_idx(m), user_data(nlohmann::json{}) {}
-KernelExecution::KernelExecution(const std::string &n, const std::string &c,
-                                 std::optional<cudaq::JitEngine> jit,
-                                 std::optional<Resources> rc, nlohmann::json &o,
-                                 std::vector<std::size_t> &m,
-                                 nlohmann::json &ud)
-    : name(n), code(c), jit(jit), resourceCounts(rc), output_names(o),
-      mapping_reorder_idx(m), user_data(ud) {}
-
-KernelExecution::~KernelExecution() = default;
-KernelExecution::KernelExecution(KernelExecution &&) noexcept = default;
-KernelExecution &
-KernelExecution::operator=(KernelExecution &&) noexcept = default;
-
-KernelExecution::KernelExecution(const KernelExecution &other)
-    : name(other.name), code(other.code), jit(other.jit),
-      resourceCounts(other.resourceCounts), output_names(other.output_names),
-      mapping_reorder_idx(other.mapping_reorder_idx),
-      user_data(other.user_data) {}
-
-KernelExecution &KernelExecution::operator=(const KernelExecution &other) {
-  return *this = KernelExecution(other);
-}
 
 void ServerHelper::parseConfigForCommonParams(const BackendConfig &config) {
   // Parse common parameters for each job and place into member variables
@@ -75,6 +41,14 @@ void ServerHelper::parseConfigForCommonParams(const BackendConfig &config) {
       this->reorderIdx[newKey] = tmp.get<std::vector<std::size_t>>();
     }
   }
+}
+
+// Out-of-line definition of the deleter for owning_ptr<ServerHelper>;
+// defined here where ServerHelper is complete so headers holding an
+// owning_ptr<ServerHelper> do not need to see the full type.
+template <>
+void opaque_deleter<ServerHelper>::operator()(ServerHelper *p) const {
+  delete p;
 }
 } // namespace cudaq
