+++ runtime/common/ServerHelper.h
@@ -11,6 +11,7 @@
 #include "CompiledModule.h"
 #include "ExecutionContext.h"
 #include "Future.h"
+#include "KernelExecution.h"
 #include "Registry.h"
 #include "Resources.h"
 #include "RuntimeTarget.h"
@@ -24,35 +25,6 @@ namespace cudaq {
 /// @brief Typedef for a mapping of key-values describing the remote server
 /// configuration.
 using BackendConfig = std::map<std::string, std::string>;
-
-/// @brief Every kernel execution has a name, compiled code representation, and
-/// (optionally) an output_names mapping showing how each Result maps back
-/// to the original program's Qubits.
-struct KernelExecution {
-  std::string name;
-  std::string code;
-  std::optional<cudaq::JitEngine> jit;
-  std::optional<Resources> resourceCounts;
-  cudaq::cudaq_json output_names;
-  std::vector<std::size_t> mapping_reorder_idx;
-  cudaq::cudaq_json user_data;
-  KernelExecution(const std::string &n, const std::string &c,
-                  std::optional<cudaq::JitEngine> jit,
-                  std::optional<Resources> rc, std::vector<std::size_t> &m);
-  KernelExecution(const std::string &n, const std::string &c,
-                  std::optional<cudaq::JitEngine> jit,
-                  std::optional<Resources> rc, nlohmann::json &o,
-                  std::vector<std::size_t> &m);
-  KernelExecution(const std::string &n, const std::string &c,
-                  std::optional<cudaq::JitEngine> jit,
-                  std::optional<Resources> rc, nlohmann::json &o,
-                  std::vector<std::size_t> &m, nlohmann::json &ud);
-  ~KernelExecution();
-  KernelExecution(const KernelExecution &other);
-  KernelExecution &operator=(const KernelExecution &other);
-  KernelExecution(KernelExecution &&) noexcept;
-  KernelExecution &operator=(KernelExecution &&) noexcept;
-};
 
 /// @brief Responses / Submissions to the Server are modeled via JSON
 using ServerMessage = nlohmann::json;
