+++ runtime/cudaq/platform.h
@@ -9,6 +9,7 @@
 #pragma once
 
 #include "cudaq/builder/kernel_builder.h"
+#include "cudaq/platform/qpu_types.h"
 #include "cudaq/platform/quantum_platform.h"
 
 namespace cudaq {
