+++ runtime/cudaq/platform/fermioniq/FermioniqRestQPU.h
@@ -1,0 +1,24 @@
+/****************************************************************-*- C++ -*-****
+ * Copyright (c) 2022 - 2026 NVIDIA Corporation & Affiliates.                  *
+ * All rights reserved.                                                        *
+ *                                                                             *
+ * This source code and the accompanying materials are made available under    *
+ * the terms of the Apache License 2.0 which accompanies this distribution.    *
+ ******************************************************************************/
+
+#pragma once
+
+#include "FermioniqBaseQPU.h"
+
+namespace cudaq {
+
+/// @brief The FermioniqRestQPU is a subtype of QPU that enables the
+/// execution of CUDA-Q kernels on the Fermioniq simulator via a REST Client.
+class FermioniqRestQPU : public FermioniqBaseQPU {
+public:
+  FermioniqRestQPU() : FermioniqBaseQPU() {}
+  FermioniqRestQPU(FermioniqRestQPU &&) = delete;
+  ~FermioniqRestQPU() override;
+};
+
+} // namespace cudaq
