+++ runtime/cudaq/platform/orca/OrcaRemoteRESTQPU.cpp
@@ -7,13 +7,10 @@
  ******************************************************************************/
 
 #include "OrcaRemoteRESTQPU.h"
+#include "common/ServerHelper.h"
 #include "cudaq/runtime/logger/logger.h"
-#include "nlohmann/json.hpp"
+#include "orca_qpu.h"
 #include "llvm/Support/Base64.h"
-
-#ifdef CUDAQ_PYTHON_EXTENSION
-extern "C" cudaq::ServerHelper *cudaq_find_server_helper(const char *name);
-#endif
 
 using namespace cudaq;
 
@@ -57,11 +54,8 @@ void cudaq::OrcaRemoteRESTQPU::setTargetBackend(const std::string &backend) {
   /// pipeline.
   // Set the qpu name
   qpuName = mutableBackend;
-#ifdef CUDAQ_PYTHON_EXTENSION
-  serverHelper.reset(cudaq_find_server_helper(qpuName.c_str()));
-#else
-  serverHelper = registry::get<ServerHelper>(qpuName);
-#endif
+  serverHelper = cudaq::owning_ptr<ServerHelper>(
+      registry::get<ServerHelper>(qpuName).release());
   serverHelper->initialize(backendConfig);
 
   // Give the server helper to the executor
@@ -102,28 +96,14 @@ KernelThunkResultType cudaq::OrcaRemoteRESTQPU::launchKernelCommon(
   return {};
 }
 
+void cudaq::OrcaRemoteRESTQPU::enqueue(cudaq::QuantumTask &task) {
+  CUDAQ_INFO("OrcaRemoteRESTQPU: Enqueue Task on QPU {}", qpu_id);
+  execution_queue->enqueue(task);
+}
+
 void cudaq::OrcaRemoteRESTQPU::launchKernel(const std::string &,
                                             const std::vector<void *> &) {
   throw std::runtime_error("launch kernel on raw args not implemented");
 }
 
-#ifdef CUDAQ_PYTHON_EXTENSION
-extern "C" void cudaq_add_qpu_node(void *node_ptr);
-
-namespace {
-struct OrcaQPURegistration {
-  cudaq::RegistryEntry<cudaq::QPU> entry;
-  cudaq::Registry<cudaq::QPU>::node node;
-  OrcaQPURegistration()
-      : entry("orca", &OrcaQPURegistration::ctorFn), node(entry) {
-    cudaq_add_qpu_node(&node);
-  }
-  static std::unique_ptr<cudaq::QPU> ctorFn() {
-    return std::make_unique<cudaq::OrcaRemoteRESTQPU>();
-  }
-};
-static OrcaQPURegistration s_orcaQPURegistration;
-} // namespace
-#else
 CUDAQ_REGISTER_TYPE(QPU, OrcaRemoteRESTQPU, orca)
-#endif
