+++ runtime/cudaq/platform/pasqal/PasqalRemoteRESTQPU.cpp
@@ -6,37 +6,8 @@
  * the terms of the Apache License 2.0 which accompanies this distribution.    *
  ******************************************************************************/
 
-#include "common/AnalogRemoteRESTQPU.h"
+#include "PasqalRemoteRESTQPU.h"
 
-namespace {
+cudaq::PasqalRemoteRESTQPU::~PasqalRemoteRESTQPU() = default;
 
-/// @brief The `PasqalRemoteRESTQPU` is a subtype of QPU that enables the
-/// execution of Analog Hamiltonian Program via a REST Client.
-class PasqalRemoteRESTQPU : public cudaq::AnalogRemoteRESTQPU {
-public:
-  PasqalRemoteRESTQPU() : AnalogRemoteRESTQPU() {}
-  PasqalRemoteRESTQPU(PasqalRemoteRESTQPU &&) = delete;
-  virtual ~PasqalRemoteRESTQPU() = default;
-};
-} // namespace
-
-#ifdef CUDAQ_PYTHON_EXTENSION
-extern "C" void cudaq_add_qpu_node(void *node_ptr);
-
-namespace {
-struct PasqalQPURegistration {
-  cudaq::RegistryEntry<cudaq::QPU> entry;
-  cudaq::Registry<cudaq::QPU>::node node;
-  PasqalQPURegistration()
-      : entry("pasqal", &PasqalQPURegistration::ctorFn), node(entry) {
-    cudaq_add_qpu_node(&node);
-  }
-  static std::unique_ptr<cudaq::QPU> ctorFn() {
-    return std::make_unique<PasqalRemoteRESTQPU>();
-  }
-};
-static PasqalQPURegistration s_pasqalQPURegistration;
-} // namespace
-#else
-CUDAQ_REGISTER_TYPE(cudaq::QPU, PasqalRemoteRESTQPU, pasqal)
-#endif
+CUDAQ_REGISTER_TYPE(cudaq::QPU, cudaq::PasqalRemoteRESTQPU, pasqal)
