+++ runtime/cudaq/platform/qpu.h
@@ -13,12 +13,7 @@
 #include "common/KernelArgs.h"
 #include "common/Registry.h"
 #include "common/ThunkInterface.h"
-#include "common/Timing.h"
-#include "cudaq/qis/execution_manager.h"
-#include "cudaq/qis/qubit_qis.h"
 #include "cudaq/remote_capabilities.h"
-#include "cudaq/runtime/logger/logger.h"
-#include "cudaq/utils/cudaq_utils.h"
 
 namespace mlir {
 class Type;
@@ -27,9 +22,14 @@ class Type;
 namespace cudaq {
 class gradient;
 class optimizer;
+class noise_model;
+class ExecutionContext;
 
-/// Expose the function that will return the current ExecutionManager
-ExecutionManager *getExecutionManager();
+// forward declare the spin_op type
+template <typename T>
+class sum_op;
+class spin_handler;
+typedef sum_op<spin_handler> spin_op;
 
 /// A CUDA-Q QPU is an abstraction on the quantum processing unit which executes
 /// quantum kernel expressions. The QPU exposes certain information about the
@@ -55,54 +55,7 @@ protected:
   /// @brief Check if the current execution context is a `spin_op` observation
   /// and perform state-preparation circuit measurement based on the `spin_op`
   /// terms.
-  void handleObservation(ExecutionContext &context) const {
-    // The reason for the 2 if checks is simply to do a flushGateQueue() before
-    // initiating the trace.
-    bool execute = context.name == "observe";
-    if (execute) {
-      ScopedTraceWithContext(cudaq::TIMING_OBSERVE,
-                             "handleObservation flushGateQueue()");
-      getExecutionManager()->flushGateQueue();
-    }
-    if (execute) {
-      ScopedTraceWithContext(cudaq::TIMING_OBSERVE,
-                             "QPU::handleObservation (after flush)");
-      double sum = 0.0;
-      if (!context.spin.has_value())
-        throw std::runtime_error("[QPU] Observe ExecutionContext specified "
-                                 "without a cudaq::spin_op.");
-
-      std::vector<cudaq::ExecutionResult> results;
-      cudaq::spin_op &H = context.spin.value();
-      assert(cudaq::spin_op::canonicalize(H) == H);
-
-      // If the backend supports the observe task, let it compute the
-      // expectation value instead of manually looping over terms, applying
-      // basis change ops, and computing <ZZ..ZZZ>
-      if (context.canHandleObserve) {
-        auto [exp, data] = cudaq::measure(H);
-        context.expectationValue = exp;
-        context.result = data;
-      } else {
-
-        // Loop over each term and compute coeff * <term>
-        for (const auto &term : H) {
-          if (term.is_identity())
-            sum += term.evaluate_coefficient().real();
-          else {
-            // This takes a longer time for the first iteration unless
-            // flushGateQueue() is called above.
-            auto [exp, data] = cudaq::measure(term);
-            results.emplace_back(data.to_map(), term.get_term_id(), exp);
-            sum += term.evaluate_coefficient().real() * exp;
-          }
-        };
-
-        context.expectationValue = sum;
-        context.result = cudaq::sample_result(sum, results);
-      }
-    }
-  }
+  void handleObservation(ExecutionContext &context) const;
 
 public:
   /// The constructor, initializes the execution queue
