+++ runtime/cudaq/platform/quera/QuEraRemoteRESTQPU.cpp
@@ -6,37 +6,8 @@
  * the terms of the Apache License 2.0 which accompanies this distribution.    *
  ******************************************************************************/
 
-#include "common/AnalogRemoteRESTQPU.h"
+#include "QuEraRemoteRESTQPU.h"
 
-namespace {
+cudaq::QuEraRemoteRESTQPU::~QuEraRemoteRESTQPU() = default;
 
-/// @brief The `QuEraRemoteRESTQPU` is a subtype of QPU that enables the
-/// execution of Analog Hamiltonian Program via a REST Client.
-class QuEraRemoteRESTQPU : public cudaq::AnalogRemoteRESTQPU {
-public:
-  QuEraRemoteRESTQPU() : AnalogRemoteRESTQPU() {}
-  QuEraRemoteRESTQPU(QuEraRemoteRESTQPU &&) = delete;
-  virtual ~QuEraRemoteRESTQPU() = default;
-};
-} // namespace
-
-#ifdef CUDAQ_PYTHON_EXTENSION
-extern "C" void cudaq_add_qpu_node(void *node_ptr);
-
-namespace {
-struct QuEraQPURegistration {
-  cudaq::RegistryEntry<cudaq::QPU> entry;
-  cudaq::Registry<cudaq::QPU>::node node;
-  QuEraQPURegistration()
-      : entry("quera", &QuEraQPURegistration::ctorFn), node(entry) {
-    cudaq_add_qpu_node(&node);
-  }
-  static std::unique_ptr<cudaq::QPU> ctorFn() {
-    return std::make_unique<QuEraRemoteRESTQPU>();
-  }
-};
-static QuEraQPURegistration s_queraQPURegistration;
-} // namespace
-#else
-CUDAQ_REGISTER_TYPE(cudaq::QPU, QuEraRemoteRESTQPU, quera)
-#endif
+CUDAQ_REGISTER_TYPE(cudaq::QPU, cudaq::QuEraRemoteRESTQPU, quera)
