+++ runtime/logger/logger.cpp
@@ -124,6 +124,9 @@ static_assert(static_cast<int>(LogLevel::info) ==
 static_assert(static_cast<int>(LogLevel::warn) ==
                   static_cast<int>(spdlog::level::warn),
               "log level enum mismatch");
+static_assert(static_cast<int>(LogLevel::error) ==
+                  static_cast<int>(spdlog::level::err),
+              "log level enum mismatch");
 bool should_log(const LogLevel logLevel) {
   return spdlog::should_log(static_cast<spdlog::level::level_enum>(logLevel));
 }
